1.1.0 - Check CHANGELOG.md

+++ main.js
@@ -2,11 +2,17 @@ const fs = require("fs-extra");
 const Database = require("better-sqlite3");
 const _ = require("lodash");
 const moment = require("moment");
+const express = require("express");
+const fastify = require("fastify")();
+const fetch = (...args) =>
+  import("node-fetch").then(({ default: fetch }) => fetch(...args));
+const app = express();
 
 module.exports = class Datarex extends Map {
   /**
    * @param {String} options.path path to database
    * @param {String} options.tableName Your tablename
+   * @param {String} options.type client or server
    * @param {Object} options.settings Settings
    * @param {Boolean} options.settings.inMemory Set to true to use cache
    * @param {Boolean} options.settings.clearCache Set to true to clear cache in intervals
@@ -22,6 +28,13 @@ module.exports = class Datarex extends Map {
     const defaultOptions = {
       path: "./Databases/index.sqlite",
       tableName: "json",
+      server: {
+        listen: "client", //default client
+        type: "client",
+        host: "127.0.0.1:433", //default 127.0.0.1
+        port: 433, // default 433
+        password: "Datarex", // default Datarex
+      },
       Intervals: {
         expiryInterval: 1000, // defaults to 1000
         clearCacheInterval: 300000, // defaults to 300000
@@ -33,7 +46,15 @@ module.exports = class Datarex extends Map {
       },
     };
     options = _.defaultsDeep(options, defaultOptions);
+    this.options = options;
     this.path = options.path;
+    this.server = options.server;
+    if (typeof options.server.port !== "number") {
+      if (typeof Parseint(options.server.port) !== "number")
+        this.server.port = parseInt(options.server.port);
+      else this.server.port = options.server.port;
+    } else this.server.port = options.server.port;
+
     this.tableName = options.tableName;
     this.settings = options.settings;
     this.Intervals = options.Intervals;
@@ -50,8 +71,24 @@ module.exports = class Datarex extends Map {
     }
     setInterval(async () => {
       try {
-        this.all().forEach((element) => {
-          if (element.VALUE.expiry <= moment().unix()) this.delete(element.KEY);
+        var stmt = this.db.prepare(
+          `SELECT * FROM ${this.tableName} WHERE KEY IS NOT NULL`
+        );
+        let resp = [];
+        for (var row of stmt.iterate()) {
+          try {
+            let VALUE = JSON.parse(row.VALUE);
+            resp.push({
+              KEY: row.KEY,
+              VALUE,
+              EXPIRY: row.EXPIRY,
+            });
+          } catch (error) {}
+        }
+        resp.forEach((element) => {
+          if (element.EXPIRY <= moment().unix()) {
+            this.removeExpiry(element.KEY);
+          }
         });
       } catch (error) {}
     }, this.Intervals.expiryInterval);
@@ -60,6 +97,42 @@ module.exports = class Datarex extends Map {
         if (this.settings.inMemory) super.clear();
       } catch (error) {}
     }, this.Intervals.clearCacheInterval);
+
+    if (this.server.type.toLowerCase() === "server") {
+      fastify.listen(this.server.port);
+      fastify.get("/Datarex", async (req, res) => {
+        const key = req?.header("key");
+        const value = req?.header("value");
+        const type = req?.header("type");
+        const tableName = req?.header("tableName");
+        const password = req?.header("password");
+        this.tableName = tableName ?? this.tableName;
+        if (password !== this.server.password)
+          return res.send({
+            error: true,
+            message: "password is incorrect",
+          });
+        if (type === undefined)
+          return res.send({
+            error: true,
+            message: "ERROR: type is undefined",
+          });
+        if (key === undefined)
+          return res.send({
+            error: true,
+            message: "key is undefined",
+          });
+        if (value === undefined)
+          return res.send({
+            error: true,
+            message: "value is undefined",
+          });
+        return res.send({
+          error: false,
+          message: (await this[type](key, value)) ?? "Action done successfully",
+        });
+      });
+    }
   }
 
   /**
@@ -67,6 +140,10 @@ module.exports = class Datarex extends Map {
    * @param {Number} value
    */
   add(key, value) {
+    this._checkMissing(key, value);
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("add", key, value);
+    this._ctine();
     if (typeof key === "number") key = key.toString();
     const get = this.get(key);
     if (get === undefined) return;
@@ -79,6 +156,9 @@ module.exports = class Datarex extends Map {
    * @returns All keys
    */
   all() {
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("all", null, null);
+    this._ctine();
     var stmt = this.db.prepare(
       `SELECT * FROM ${this.tableName} WHERE KEY IS NOT NULL`
     );
@@ -97,7 +177,10 @@ module.exports = class Datarex extends Map {
   /**
    *
    */
-  async backup(options = {}) {
+  async backUp(options = {}) {
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("backup", options, null);
+    this._ctine();
     if (!options.name)
       options.name ===
         `backup-${new Date().getDate()}-${new Date().getMonth()}-${new Date().getFullYear()}`;
@@ -116,8 +199,8 @@ module.exports = class Datarex extends Map {
     )
       throw TypeError(`
         ${pico.red(
-          `Backup database names cannot include there special characters: `
-        )}/\\?*":<>`);
+          `Backup database names cannot include there special characters:`
+        )} /\\?*":<>`);
 
     const dbName = options.path + options.name;
 
@@ -127,18 +210,28 @@ module.exports = class Datarex extends Map {
    * Clears the cache
    */
   clearCache() {
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("clearCache", null, null);
+    this._ctine();
     super.clear();
   }
   /**
    * closes the database
    */
   close() {
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("close", null, null);
+    this._ctine();
     this.db.close();
   }
   /**
    *@param {String} key
    */
   delete(key) {
+    this._checkMissing(key, null);
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("delete", key, null);
+    this._ctine();
     if (typeof key === "number") key = key.toString();
     if (this.settings.inMemory) super.delete(key);
     this.db.prepare(`DELETE FROM ${this.tableName} WHERE KEY = (?)`).run(key);
@@ -149,6 +242,9 @@ module.exports = class Datarex extends Map {
    * @param {any} value
    */
   deleteAll() {
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("deleteAll", null, null);
+    this._ctine();
     if (this.settings.inMemory) super.clear();
     this.db.prepare(`DELETE FROM ${this.tableName}`).run();
   }
@@ -159,6 +255,10 @@ module.exports = class Datarex extends Map {
    * divides 2 numbers and set it
    */
   divide(key, value) {
+    this._checkMissing(key, value);
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("divide", key, value);
+    this._ctine();
     if (typeof key === "number") key = key.toString();
     const get = this.get(key);
 
@@ -173,7 +273,9 @@ module.exports = class Datarex extends Map {
    * @returns {any}
    */
   get(key) {
-    const okey = key;
+    this._checkMissing(key, null);
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("get", key, null);
     this._ctine();
     if (typeof key === "number") key = key.toString();
     let target;
@@ -213,6 +315,10 @@ module.exports = class Datarex extends Map {
    * @returns {Boolean}
    */
   has(key) {
+    this._checkMissing(key, null);
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("has", key, null);
+    this._ctine();
     if (typeof key === "number") key = key.toString();
     return Boolean(
       this.db
@@ -227,6 +333,10 @@ module.exports = class Datarex extends Map {
    * multiplies 2 numbers and set it
    */
   multiply(key, value) {
+    this._checkMissing(key, value);
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("multiply", key, value);
+    this._ctine();
     if (typeof key === "number") key = key.toString();
     const get = this.get(key);
 
@@ -241,6 +351,10 @@ module.exports = class Datarex extends Map {
    * @param {any} value
    */
   push(key, value) {
+    this._checkMissing(key, value);
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("push", key, value);
+    this._ctine();
     if (typeof key === "number") key = key.toString();
     const get = this.get(key);
     if (typeof get !== "object") throw Error(`${key} is not an array`);
@@ -251,6 +365,10 @@ module.exports = class Datarex extends Map {
    * @param {Function} value
    */
   removeArrVal(key, value) {
+    this._checkMissing(key, value);
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("removeArrVal", key, value);
+    this._ctine();
     const data = this.get(key);
     if (!_.isArray(data)) throw Error(`${key} is not an array`);
     const criteria = isFunction(value) ? value : (val) => value === val;
@@ -263,8 +381,32 @@ module.exports = class Datarex extends Map {
   /**
    * @param {String} key
    */
+  removeExpiry(key) {
+    this._checkMissing(key, null);
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("removeExpiry", key, null);
+    let value = this.get(key);
+    if (
+      Boolean(
+        this.db
+          .prepare(`SELECT EXPIRY FROM ${this.tableName} WHERE KEY = (?)`)
+          .get(key)
+      )
+    )
+      this.db
+        .prepare(
+          `UPDATE ${this.tableName} SET KEY = (?), VALUE = (?), EXPIRY = (?) WHERE KEY = (?)`
+        )
+        .run(key, JSON.stringify(value), "null", key);
+  }
+  /**
+   * @param {String} key
+   */
   removeObjVal(key) {
-    let Fkey = key;
+    this._checkMissing(key, null);
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("removeObjVal", key, null);
+    this._ctine();
     let Gkey;
     if (key.includes(".")) {
       key = key.split(".");
@@ -280,7 +422,10 @@ module.exports = class Datarex extends Map {
    * @param {String} KEY
    * @param {any} value
    */
-  set(key, value) {
+  set(key, value, options = { EXPIRY: "null" }) {
+    this._checkMissing(key, value);
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("set", key, value);
     this._ctine();
     if (typeof key === "number") key = key.toString();
     let target;
@@ -295,8 +440,10 @@ module.exports = class Datarex extends Map {
 
     if (!Boolean(fetched)) {
       this.db
-        .prepare(`INSERT INTO ${this.tableName} (KEY , VALUE) VALUES( ? , ? )`)
-        .run(key, "{}");
+        .prepare(
+          `INSERT INTO ${this.tableName} (KEY , VALUE, EXPIRY ) VALUES( ? , ? , ? )`
+        )
+        .run(key, "{}", `${options.EXPIRY}`);
       fetched = this.db
         .prepare(`SELECT * FROM ${this.tableName} WHERE KEY = (?)`)
         .get(key);
@@ -309,21 +456,22 @@ module.exports = class Datarex extends Map {
 
     this.db
       .prepare(
-        `UPDATE ${this.tableName} SET KEY = (?), VALUE = (?) WHERE KEY = (?)`
+        `UPDATE ${this.tableName} SET KEY = (?), VALUE = (?), EXPIRY = (?) WHERE KEY = (?)`
       )
-      .run(key, value, key);
+      .run(key, value, options.EXPIRY, key);
 
     if (this.settings.inMemory) super.set(key, value);
   }
 
   /**
    * @param {String} key
-   * @param {String} value
    * @param {String} timeString
    */
   setExpiry(key, timeString) {
+    this._checkMissing(key, timeString);
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("setExpiry", key, timeString);
     this._ctine();
-    let target;
     if (key && key.includes(".")) {
       let unparsed = key.split(".");
       key = unparsed.shift();
@@ -331,11 +479,12 @@ module.exports = class Datarex extends Map {
     }
     const Nkey = key;
     key = this.get(key);
+    console.log(key);
     if (!key) throw Error(`${key} does not exists`);
-    const time = this._parseTime(timeString);
-    key.expiry = time / 1000 + moment().unix();
-    key.now = moment().unix();
-    this.set(Nkey, key);
+    const time = this._parseTime(timeString) ?? this._parseTime("10m");
+    this.set(`${Nkey}.expiry`, key, {
+      EXPIRY: `${time / 1000 + moment().unix()}`,
+    });
   }
 
   /**
@@ -343,6 +492,9 @@ module.exports = class Datarex extends Map {
    *
    */
   setMany(array) {
+    this._checkMissing(key, null);
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("setmany", array, null);
     this._ctine();
     if (!_.isArray(array)) throw Error(`${array} is not an array`);
     array.forEach((element, index) => {
@@ -357,6 +509,9 @@ module.exports = class Datarex extends Map {
    * @param {Number} value
    */
   subtract(key, value) {
+    this._checkMissing(key, value);
+    if (this.server.listen.toLowerCase() === "server")
+      return this._Server("subtract", key, value);
     this._ctine();
     if (typeof key === "number") key = key.toString();
     const get = this.get(key);
@@ -380,9 +535,36 @@ module.exports = class Datarex extends Map {
   _ctine() {
     this.db
       .prepare(
-        `CREATE TABLE IF NOT EXISTS ${this.tableName} (KEY TEXT, VALUE TEXT)`
+        `CREATE TABLE IF NOT EXISTS ${this.tableName} (KEY TEXT, VALUE TEXT, EXPIRY TEXT)`
       )
       .run();
+  }
+  /**
+   * @private
+   */
+  _checkMissing(key, value) {
+    if (key === undefined) throw new TypeError(`${key} is not provided`);
+    if (value === undefined) throw new TypeError(`${value} is not provided`);
+  }
+
+  async _Server(type, key, value) {
+    if (this.server.listen.toLowerCase() === "client") return;
+
+    return await fastify
+      .post(`${this.server.host}/Datarex`, {
+        headers: {
+          type: type,
+          key: key,
+          value: value,
+          tableName: this.tableName,
+          password: this.server.password,
+        },
+      })
+      .then(async (response, error) => {
+        if (error) return error;
+        if (response.json()) return await response.json();
+      })
+      .catch((error) => console.log(error));
   }
 
   /**
